major (#252): Finished phone optimisation for editor view
WIP! Presentation preview not phone optimised
