+++ solardoc/frontend/src/assets/core/_button.scss
@@ -84,7 +84,7 @@ button.editor-button {
   font-size: $editor-button-font-size;
   text-decoration: none;
   background: transparent;
-  padding: 0.5rem 1rem;
+  padding: 0.5rem 0.8rem;
   margin: 0;
 
   &:hover {
