+++ solardoc/frontend/src/assets/core/_var.scss
@@ -302,7 +302,8 @@ $highlighted-button-font-size: var(--highlighted-button-font-size);
 $window-2xlarge: 1500px;
 $window-xlarge: 1280px;
 $window-large: 1140px;
-$window-xmedium: 1024px;
+$window-2xmedium: 1024px;
+$window-xmedium: 960px;
 $window-medium: 768px;
 $window-small: 640px;
 $window-xsmall: 480px;
